requirements.txt
wheel
Django==4.0.1
django-crispy-forms==1.13.0
djangorestframework==3.13.1
pytest==6.2.5
pyodbc==4.0.34